refactor(app): move caching code to subfolder
