+++ lib/app/app.dart
@@ -11,11 +11,15 @@ export 'package:pedantic/pedantic.dart';
 export 'package:schulcloud/brand/brand.dart';
 export 'package:time_machine/time_machine.dart' show Instant, LocalDate, Period;
 
+export 'account/avatar.dart';
 export 'banner/service.dart';
+export 'caching/exception.dart';
+export 'caching/pages/empty_state.dart';
+export 'caching/utils.dart';
+export 'caching/widgets/error.dart';
 export 'data.dart';
 export 'datetime_utils.dart';
 export 'error_reporting.dart';
-export 'exception.dart';
 export 'hive.dart';
 export 'logger.dart';
 export 'routing.dart' show FancyRoute, FancyRouteBuilder;
@@ -29,9 +33,6 @@ export 'sort_filter/sort_filter.dart';
 export 'sort_filter/sorting.dart';
 export 'utils.dart';
 export 'widgets/app_bar.dart';
-export 'widgets/cache_utils.dart';
-export 'widgets/empty_state.dart';
-export 'widgets/error_widgets.dart';
 export 'widgets/fade_in.dart';
 export 'widgets/form.dart';
 export 'widgets/scaffold.dart';
