+++ lib/app/sort_filter/filtering.dart
@@ -6,9 +6,9 @@ import 'package:intl/intl.dart';
 import 'package:time_machine/time_machine.dart';
 import 'package:time_machine/time_machine_text_patterns.dart';
 
+import '../caching/utils.dart';
 import '../datetime_utils.dart';
 import '../utils.dart';
-import '../widgets/cache_utils.dart';
 import 'sort_filter.dart';
 
 typedef Test<T, D> = bool Function(T item, D data);
