+++ lib/assignment/pages/assignments_page.dart
@@ -86,7 +86,7 @@ class _AssignmentsPageState extends State<AssignmentsPage>
             title: Text(s.assignment),
             actions: <Widget>[SortFilterIconButton(showSortFilterSheet)],
           ),
-          emptyStateBuilder: (context) => EmptyStateScreen(
+          emptyStateBuilder: (context) => EmptyStatePage(
             text: context.s.assignment_assignmentsPage_empty,
           ),
           sortFilterSelection: sortFilterSelection,
