+++ lib/news/widgets/news_screen.dart
@@ -13,9 +13,8 @@ class NewsScreen extends StatelessWidget {
         collection: services.storage.root.news,
         builder: handleLoadingErrorRefreshEmpty(
           appBar: FancyAppBar(title: Text(context.s.news)),
-          emptyStateBuilder: (context) => EmptyStateScreen(
-            text: context.s.news_empty,
-          ),
+          emptyStateBuilder: (context) =>
+              EmptyStatePage(text: context.s.news_empty),
           builder: (context, unsortedArticles, isFetching) {
             final articles = unsortedArticles.toList()
               ..sort((a1, a2) => a2.publishedAt.compareTo(a1.publishedAt));
